split up passwords module
