+++ src/klein/storage/passwords/__init__.py
@@ -1,17 +1,9 @@
-from ._scrypt import (
-    InvalidPasswordRecord,
-    KleinV1PasswordEngine,
-    PasswordEngine,
-    engineForTesting,
-    hashUpgradeCount,
-)
+from ._interfaces import PasswordEngine
+from ._scrypt import InvalidPasswordRecord, KleinV1PasswordEngine
 
 
 __all__ = [
     "InvalidPasswordRecord",
     "KleinV1PasswordEngine",
     "PasswordEngine",
-    # testing
-    "engineForTesting",
-    "hashUpgradeCount",
 ]
