+++ src/klein/storage/passwords/testing.py
@@ -1,0 +1,8 @@
+# -*- test-case-name: klein.storage.passwords.test.test_passwords -*-
+from ._testing import engineForTesting, hashUpgradeCount
+
+
+__all__ = [
+    "engineForTesting",
+    "hashUpgradeCount",
+]
